pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "storymap"
version = "0.1.0"
description = "Generate user story maps from markdown"
requires-python = ">=3.11"
dependencies = [
    "markdown-it-py>=3.0",
    "weasyprint>=61.0",
    "click>=8.0",
]

[project.scripts]
storymap = "storymap.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
